+++ old_examples/Standards/full_32.py
