+++ old_examples/Standards/full_32_fast.py
