+++ old_examples/Standards/full_survey.py
