+++ old_examples/Standards/full_zenithmask.py
