+++ old_examples/Standards/generate_dd.py
