+++ old_examples/Standards/one_filter_dither.py
