+++ old_examples/Standards/one_filter_pairs.py
