+++ old_examples/Standards/one_filter_rolling.py
