+++ old_examples/Standards/one_filter_unmasked.py
