+++ old_examples/Standards/six_filter.py
