+++ old_examples/block_pairs/mix_filters.py
