+++ old_examples/block_pairs/same_filters.py
