+++ old_examples/comcam_example.py
