+++ old_examples/pairs_example.py
