+++ old_examples/six_w_pairs.py
