+++ old_examples/two_filters_example.py
